Move sub-classes of UndoCommand into a different package
