+++ src/main/java/taskle/logic/commands/UndoCommand.java
@@ -4,6 +4,12 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
+import taskle.logic.commands.undo.UndoAddCommand;
+import taskle.logic.commands.undo.UndoClearCommand;
+import taskle.logic.commands.undo.UndoDoneCommand;
+import taskle.logic.commands.undo.UndoEditCommand;
+import taskle.logic.commands.undo.UndoRemoveCommand;
+import taskle.logic.commands.undo.UndoRescheduleCommand;
 import taskle.logic.history.History;
 import taskle.model.Model;
