+++ src/main/java/taskle/logic/commands/undo/UndoRemoveCommand.java
@@ -1,5 +1,9 @@
-package taskle.logic.commands;
+package taskle.logic.commands.undo;
 
+import taskle.logic.commands.Command;
+import taskle.logic.commands.CommandResult;
+import taskle.logic.commands.RemoveCommand;
+import taskle.logic.commands.UndoCommand;
 import taskle.model.Model;
 import taskle.model.task.Task;
 import taskle.model.task.UniqueTaskList;
