+++ src/main/java/taskle/logic/commands/undo/UndoRescheduleCommand.java
@@ -1,9 +1,13 @@
-package taskle.logic.commands;
+package taskle.logic.commands.undo;
 
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
 
+import taskle.logic.commands.Command;
+import taskle.logic.commands.CommandResult;
+import taskle.logic.commands.RescheduleCommand;
+import taskle.logic.commands.UndoCommand;
 import taskle.model.Model;
 import taskle.model.task.DeadlineTask;
 import taskle.model.task.EventTask;
